🚚 resume pydantic 스키마 분리
