+++ resume/schemas/submission_schemas.py
@@ -4,104 +4,8 @@ from uuid import UUID
 
 from pydantic import BaseModel
 
-from job_posting.schemas import JobPostingListModel
-from user.models import UserInfo
-from user.schemas import UserInfoModel
+from resume.schemas.common_schemas import CareerInfoModel, CertificationInfoModel
 from utils.schemas import MY_CONFIG
-
-# ------------------------
-# Career (경력 정보)
-# ------------------------
-
-
-class CareerInfoModel(BaseModel):
-    model_config = MY_CONFIG
-
-    company_name: str
-    position: str
-    employment_period_start: date
-    employment_period_end: Optional[date] = None
-
-
-# ------------------------
-# Certification (자격증)
-# ------------------------
-
-
-class CertificationInfoModel(BaseModel):
-    model_config = MY_CONFIG
-
-    certification_name: str
-    issuing_organization: str
-    date_acquired: date
-
-
-# ------------------------
-# Resume (이력서)
-# ------------------------
-
-
-class ResumeCreateModel(BaseModel):
-    """
-    이력서 생성 model
-    """
-
-    model_config = MY_CONFIG
-    job_category: str = ""
-    resume_title: str
-    education_level: str
-    school_name: str
-    education_state: str
-    introduce: str
-
-    career_list: Optional[List[CareerInfoModel]] = None
-    certification_list: Optional[List[CertificationInfoModel]] = None
-
-
-class ResumeUpdateModel(BaseModel):
-    """
-    이력서 수정 model
-    """
-
-    resume_id: UUID
-    job_category: Optional[str] = None
-    resume_title: Optional[str] = None
-    education_level: Optional[str] = None
-    school_name: Optional[str] = None
-    education_state: Optional[str] = None
-    introduce: Optional[str] = None
-    career_list: Optional[List[CareerInfoModel]] = None
-    certification_list: Optional[List[CertificationInfoModel]] = None
-
-
-class ResumeListOutputModel(BaseModel):
-    """
-    이력서 리스트 조회
-    """
-
-    model_config = MY_CONFIG
-    resume_id: UUID
-    resume_title: str
-    job_category: str
-    updated_at: date
-
-
-class ResumeOutputModel(BaseModel):
-    """
-    이력서
-    """
-
-    model_config = MY_CONFIG
-    resume_id: UUID
-    job_category: str = ""
-    resume_title: str
-    education_level: str
-    school_name: str
-    education_state: str
-    introduce: str
-    user: UserInfoModel
-    career_list: Optional[List[CareerInfoModel]]
-    certification_list: Optional[List[CertificationInfoModel]]
 
 
 class SubmissionOutputModel(BaseModel):
@@ -276,54 +180,6 @@ class SubmissionCompanyGetListOutputModel(BaseModel):
 # ------------------------
 # 응답 모델
 # ------------------------
-
-
-class ResumeResponseModel(BaseModel):
-    message: str
-    resume: ResumeOutputModel
-
-
-class MyResume(BaseModel):
-    """
-    이력서 상세보기 시 id, title
-    """
-
-    model_config = MY_CONFIG
-    resume_id: UUID
-    resume_title: str
-
-
-class MyResumeListOutput(BaseModel):
-    """
-    이력서 상세보기 시 id, title 리스트
-    """
-
-    model_config = MY_CONFIG
-    resume_id: UUID
-    job_category: str = ""
-    resume_title: str
-    education_level: str
-    school_name: str
-    education_state: str
-    introduce: str
-    user: UserInfoModel
-    career_list: Optional[List[CareerInfoModel]]
-    certification_list: Optional[List[CertificationInfoModel]]
-    resume_list: list[MyResume]
-
-
-class MyResumeMixinResponse(BaseModel):
-    """
-    resume 상세 조회 시 유저의 resume 목록(id, title)포함 모델
-    """
-
-    message: str
-    resume: MyResumeListOutput
-
-
-class ResumeListResponseModel(BaseModel):
-    message: str
-    resume_list: List[ResumeListOutputModel]
 
 
 class SubmissionListResponseModel(BaseModel):
