+++ utils/mock_test_data.py
@@ -4,7 +4,7 @@ from django.utils import timezone
 
 from job_posting.models import JobPosting
 from resume.models import CareerInfo, Certification, Resume, Submission
-from resume.schemas import CareerInfoModel, CertificationInfoModel
+from resume.schemas.common_schemas import CareerInfoModel, CertificationInfoModel
 from user.models import CommonUser, CompanyInfo, UserInfo
